(chore, bug, enhancement) Refactored server file:
* Broke server.js into the following files:
    * server.js, server/utils/middleware.js, routes.js
    * Refactored app.js and changed name from app.js => index.js
