+++ src/reducers/OfferReducers.jsx
@@ -3,7 +3,7 @@ import { CHANGE_OFFER } from '../actions/Offer_Actions';
 // Auth reducer, starting state sets auth based on a token in local storage
 // TODO: A Util to check if the token is expired
 export default function(state = {
-  offer: false
+  offer: false,
 }, action) {
   switch (action.type) {
     case CHANGE_OFFER:
@@ -11,6 +11,6 @@ export default function(state = {
         offer: action.offer,
       });
     default:
-      return state,
+      return state;
   }
 }
